First Step to Individual LED Map
Currently there is a map of the keys on the keyboard and the
corresponding LEDs attached to those positions. This sets up the basis
for having individual addressable LEDs (tied back to the keys or
positions on the keyboard).

TODO: The mapping just copies back the KeyInfo object. A new model
should be created that just identifies the LED and has a ref to the
KeyInfo model.